Moves `ConfigProvider` into subnamespace
Moves the `ConfigProvider` into the subnamespace
`Zend\Expressive\Router\ZendRouter` in order to prevent naming
collisions.
